Add fetcher Todo: merge dbclient
