+++ internal/cloud/gcpcln/client.go
@@ -11,8 +11,6 @@ import (
 	"context"
 	"encoding/csv"
 	"fmt"
-	"log"
-	"time"
 
 	"cloud.google.com/go/storage"
 	"github.com/pavlov-tony/xproject/pkg/cloud/gcpparser"
@@ -130,33 +128,4 @@ func (c *Client) MakeReports(objs Objects) (reps Reports, err error) {
 	}
 
 	return reps, nil
-}
-
-// Fetch periodically fetches data into db from GCP
-func (c *Client) Fetch(bktName, prefix string, dt time.Duration) {
-	for {
-		// pull all csv objects from bucket
-		objs, err := c.CsvObjsList(bktName, prefix)
-		if err != nil {
-			log.Fatalf("in fetch CsvObjList: %v", err)
-		}
-
-		// select last report from db with its creation time
-		// TODO: use pgcln here to select last report from db
-
-		// filter CsvObjList, save only fresh objects
-		// TODO: objs := objs.after(lastReport.Object.Created)
-
-		// make Reports from object content
-		reps, err := c.MakeReports(objs)
-		_ = reps
-		if err != nil {
-			log.Fatalf("in fetch MakeReports: %v", err)
-		}
-
-		// write Reports into db
-		// TODO: use pgcln here to write parsed csv into db
-
-		time.Sleep(dt * time.Hour)
-	}
 }
